# openapi.yaml
openapi: 3.0.3
info:
  title: AdrenRush API
  version: "1.0.0"
  description: >
    REST API сайта рейтинга энергетиков **AdrenRush**.

    Аутентификация — JWT в заголовке `Authorization: Bearer <token>`.
    Токен выдаётся при регистрации/входе (`/api/auth/register`, `/api/auth/login`).

    Уровни доступа (см. `SecurityConfig`):
    - `GET /api/drinks/**`, `GET /api/users/**`, `GET /media/**`, `/api/auth/**` — без авторизации;
    - `/api/admin/**` — только роль `ADMIN`;
    - изменяющие операции над напитками/фото — требуют токен и проверяются на роль `ADMIN` в контроллере (иначе `403`);
    - остальные эндпоинты требуют валидный токен.

    Ошибки бизнес-логики возвращаются как `{"error": "..."}` с соответствующим HTTP-статусом.
servers:
  - url: http://localhost:8080
    description: Локальная разработка
  - url: https://adrenrush.ru
    description: Боевой сервер
...
security:
  - bearerAuth: []
paths:
...
  /api/drinks:
    get:
      tags: [Drinks]
      summary: Список энергетиков (по убыванию оценки)
      description: Краткие карточки для главной (без полной галереи).
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: Список энергетиков
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DrinkResponseDto'
    post:
      tags: [Drinks]
      summary: Создать карточку энергетика (только админ)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDrinkRequest'
      responses:
        '200':
          description: Созданный энергетик
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrinkResponseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/drinks/{id}:
    get:
      tags: [Drinks]
      summary: Карточка энергетика с галереей
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Полная карточка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrinkResponseDto'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Drinks]
      summary: Редактировать энергетик (только админ)
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDrinkRequest'
      responses:
        '200':
          description: Обновлённый энергетик
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrinkResponseDto'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Drinks]
      summary: Удалить энергетик (только админ)
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Удалён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/drinks/{id}/photos:
    post:
      tags: [Drinks]
      summary: Добавить фото в галерею файлом (только админ)
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Энергетик с обновлённой галереей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrinkResponseDto'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/drinks/{id}/photos/url:
    post:
      tags: [Drinks]
      summary: Добавить фото по ссылке (скачивается в хранилище, только админ)
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPhotoUrlRequest'
      responses:
        '200':
          description: Энергетик с обновлённой галереей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrinkResponseDto'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/drinks/{id}/photos/order:
    put:
      tags: [Drinks]
      summary: Изменить порядок фотографий (первое = обложка, только админ)
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderPhotosRequest'
      responses:
        '200':
          description: Энергетик с переупорядоченной галереей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrinkResponseDto'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/drinks/{id}/photos/{photoId}:
    delete:
      tags: [Drinks]
      summary: Удалить фото из галереи (только админ)
      parameters:
        - $ref: '#/components/parameters/PathId'
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Энергетик с обновлённой галереей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrinkResponseDto'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/drinks/{id}/cover:
    put:
      tags: [Drinks]
      summary: Настроить кадрирование обложки (только админ)
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCoverRequest'
      responses:
        '200':
          description: Энергетик с обновлённым кадрированием
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrinkResponseDto'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/drinks/media/optimize:
    post:
      tags: [Drinks]
      summary: Оптимизировать медиа (только админ)
      description: >
        Внешние картинки скачиваются в наше хранилище, для фото без превью оно достраивается.
        Сетевые загрузки могут занять время — ответ приходит по завершении.
      responses:
        '200':
          description: Итог оптимизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaOptimizeResult'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/drinks/parse/sources:
    get:
      tags: [Drinks]
      summary: Источники с авто-парсером каталога (только админ)
      description: >
        Monster берётся не с официального сайта (он за Cloudflare), а из каталога магазина
        worldsweet.ru — источник «Monster (WorldSweet)».
      responses:
        '200':
          description: Список источников
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example: ["Adrenaline Rush", "Red Bull", "Monster (WorldSweet)"]
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/drinks/parse:
    post:
      tags: [Drinks]
      summary: Запустить авто-парсеры для выбранных брендов (только админ)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParseRequest'
      responses:
        '200':
          description: Сколько карточек создано/обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ────────────────────────────── REVIEWS ───────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Требуется авторизация
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Недостаточно прав (или аккаунт заблокирован)
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/Error'
              - $ref: '#/components/schemas/BannedError'
    NotFound:
      description: Объект не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
      example: { error: "Сообщение об ошибке" }
...
    BannedError:
      type: object
      description: Возвращается JWT-фильтром, если аккаунт заблокирован
      properties:
        error:
          type: string
          example: Аккаунт заблокирован
        reason:
          type: string
        bannedUntil:
          type: string
          description: ISO-время до которого действует бан, либо пустая строка (бан навсегда)
      required: [error]
...
    StatusResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
      required: [status]
...
    DrinkResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        brand:
          type: string
          nullable: true
        slug:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
          description: Обложка; в кратком виде — заданная, в полном — url первого фото
        coverFitCard:
          type: string
          nullable: true
        coverPosCard:
          type: string
          nullable: true
        coverFitModal:
          type: string
          nullable: true
        coverPosModal:
          type: string
          nullable: true
        averageRating:
          type: number
          format: double
        reviewCount:
          type: integer
        ratingDistribution:
          type: object
          nullable: true
          description: 'Распределение оценок: балл (1–10) → количество'
          additionalProperties:
            type: integer
          example: { "8": 3, "9": 5, "10": 12 }
        photos:
          type: array
          nullable: true
          description: Полная галерея (только в карточке по id)
          items:
            $ref: '#/components/schemas/PhotoDto'
    PhotoDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        url:
          type: string
        thumbUrl:
          type: string
          nullable: true
        source:
          $ref: '#/components/schemas/PhotoSource'
        uploadedById:
          type: integer
          format: int64
          nullable: true

    # ── Отзывы ──
...
    ParseResult:
      type: object
      properties:
        created:
          type: integer
        updated:
          type: integer
    MediaOptimizeResult:
      type: object
      properties:
        downloaded:
          type: integer
        thumbnailed:
          type: integer
        skipped:
          type: integer
        failed:
          type: integer
...
    PhotoSource:
      type: string
      enum: [PARSED, USER]
...
    CreateDrinkRequest:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        description:
          type: string
        coverUrl:
          type: string
      required: [name]
    UpdateDrinkRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    AddPhotoUrlRequest:
      type: object
      properties:
        url:
          type: string
      required: [url]
    ReorderPhotosRequest:
      type: object
      properties:
        order:
          type: array
          items:
            type: integer
            format: int64
          description: Идентификаторы фото в желаемом порядке (первое = обложка)
      required: [order]
    UpdateCoverRequest:
      type: object
      description: Параметры кадрирования обложки (CSS object-fit / object-position)
      properties:
        coverFitCard:
          type: string
        coverPosCard:
          type: string
        coverFitModal:
          type: string
        coverPosModal:
          type: string
    ParseRequest:
      type: object
      properties:
        brands:
          type: array
          items:
            type: string
          description: Источники (бренды/каталоги) из /api/drinks/parse/sources
        reparse:
          type: boolean
          default: false
          description: false — только новые карточки; true — обновить и существующие
...